Add back ability to prevent deleting packaged and custom scripts after running them
Fixes OctopusDeploy/issues#2213

+++ source/Calamari/Deployment/Conventions/PackagedScriptConvention.cs
@@ -17,7 +17,10 @@ namespace Calamari.Deployment.Conventions
         public void Install(RunningDeployment deployment)
         {
             RunScripts(deployment);
-            DeleteScripts(deployment);
+            if (deployment.Variables.GetFlag(SpecialVariables.DeleteScriptsOnCleanup, true))
+            {
+                DeleteScripts(deployment);
+            }
         }
     }
 }
